Make InlineCommentVisitor support nested modules defined with w/ ::
For these modules, `visit_{class,module}_node` will be called only once, and `node.name` returns only the base name of the module.
Because of that, there were missing signature comments in cases like `it "inline signatures will be written for methods in nested modules defined w/ ::"`.
This change fixes that problem by making the visitor recognize the full path of such kinds of modules.